third_party/blink/web_tests/svg/dynamic-updates/SVGFEDropShadowElement-svgdom-shadow-color-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'flood-color' property of the SVGFEFloodElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.defaultView.getComputedStyle(dropShadowElement).getPropertyValue('flood-color') is "rgb(0, 0, 0)"
PASS document.defaultView.getComputedStyle(dropShadowElement).getPropertyValue('flood-color') is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
